+++ test/rich-text/commands/index.test.ts
@@ -7,7 +7,6 @@ import {
     toggleHeadingLevel,
     toggleTagLinkCommand,
     toggleWrapIn,
-    toggleCodeBlock,
 } from "../../../src/rich-text/commands";
 import {
     applyNodeSelection,
@@ -17,8 +16,6 @@ import {
 } from "../test-helpers";
 import { toggleMark } from "prosemirror-commands";
 import { MarkType } from "prosemirror-model";
-import { schema as basicSchema } from "prosemirror-schema-basic";
-import { doc, p, code_block, br } from "prosemirror-test-builder";
 import { EditorView } from "prosemirror-view";
 
 function getEndOfNode(state: EditorState, nodePos: number) {
@@ -1242,94 +1239,4 @@ describe("commands", () => {
             });
         });
     });
-
-    describe("toggleCodeBlock command", () => {
-        test("turns an empty paragraph into an empty code block", () => {
-            // Start with a doc of one empty paragraph, cursor inside it
-            const startDoc = doc(p("<cursor>"));
-
-            let state = EditorState.create({
-                doc: startDoc,
-                schema: basicSchema,
-            });
-
-            // Apply the command
-            state = applyCommand(toggleCodeBlock(), state);
-
-            // The resulting doc should be just a codeBlock node, empty
-            // (In the basic schema, codeBlock is "```" style, but let's just match structure).
-            const expectedDoc = doc(code_block());
-
-            expect(state.doc.toJSON()).toEqual(expectedDoc.toJSON());
-        });
-
-        test("turns a paragraph with text into a code block (no selection, just cursor)", () => {
-            // Start: paragraph with text, cursor in the middle
-            const startDoc = doc(p("ab<cursor>cd"));
-
-            let state = EditorState.create({
-                doc: startDoc,
-                schema: basicSchema,
-            });
-
-            state = applyCommand(toggleCodeBlock(), state);
-
-            // The entire paragraph block should become a single codeBlock node.
-            // The text is "abcd". The command merges it into one line (since there's
-            // only one block, no line breaks).
-            const expectedDoc = doc(code_block("abcd"));
-
-            expect(state.doc.toJSON()).toEqual(expectedDoc.toJSON());
-        });
-
-        test("turns a multi-line code block into a paragraph with hard breaks", () => {
-            // A codeBlock with two lines: "abc\ndef"
-            // We'll put the cursor anywhere in that block.
-            const startDoc = doc(code_block("abc\ndef<cursor>"));
-
-            let state = EditorState.create({
-                doc: startDoc,
-                schema: basicSchema,
-            });
-
-            state = applyCommand(toggleCodeBlock(), state);
-
-            // We should get a single paragraph with "abc", then a hard_break, then "def".
-            // Using test-builder, we can represent a hard break with `br()`.
-            const expectedDoc = doc(p("abc", br(), "def"));
-
-            expect(state.doc.toJSON()).toEqual(expectedDoc.toJSON());
-        });
-
-        test("turns a paragraph with a hard_break into a multi-line code block", () => {
-            // A single paragraph, "abc<break>def"
-            const startDoc = doc(p("abc", br(), "def<cursor>"));
-
-            let state = EditorState.create({
-                doc: startDoc,
-                schema: basicSchema,
-            });
-
-            state = applyCommand(toggleCodeBlock(), state);
-
-            // This should yield one codeBlock node with "abc\ndef".
-            const expectedDoc = doc(code_block("abc\ndef"));
-
-            expect(state.doc.toJSON()).toEqual(expectedDoc.toJSON());
-        });
-    });
 });
-
-type Command = (
-    state: EditorState,
-    dispatch?: (tr: Transaction) => void,
-    view?: EditorView
-) => boolean;
-
-function applyCommand(command: Command, state: EditorState): EditorState {
-    let newState = state;
-    command(state, (tr: Transaction) => {
-        newState = state.apply(tr);
-    });
-    return newState;
-}
